Enhance Dapr support in Aspire framework
- Added project reference to `CommunityToolkit.Aspire.Hosting.Dapr.csproj` in the Azure Dapr project file.
- Updated `IDistributedApplicationBuilderExtensions.cs` with documentation for Dapr-related extension methods.
- Introduced `CommandLineBuilder.cs` for building Dapr CLI command line arguments.
- Modified and added classes for Dapr components, improving configuration and management.
- Refactored `DaprDistributedApplicationLifecycleHook` to manage Dapr sidecar lifecycles and on-demand components.
- Introduced new annotations and options classes for better Dapr component configuration.
- Improved organization and clarity of Dapr-related classes with additional comments and documentation.
