+++ src/CommunityToolkit.Aspire.Hosting.Azure.Dapr/CommunityToolkit.Aspire.Hosting.Azure.Dapr.csproj
@@ -5,15 +5,6 @@
     <Description>Azure Dapr support for .NET Aspire.</Description>
   </PropertyGroup>
 
-  
-
-<ItemGroup>
-  <Compile Include="$(SharedDir)\Dapr\Core\*.cs">
-    <Link>%(Filename)%(Extension)</Link>
-  </Compile>
-</ItemGroup>
-
-
   <ItemGroup>
     <Compile Include="$(SharedDir)\Utf8JsonWriterExtensions.cs" Link="Utils\Utf8JsonWriterExtensions.cs" />
   </ItemGroup>
@@ -22,6 +13,10 @@
     <PackageReference Include="Aspire.Hosting" />
     <PackageReference Include="Aspire.Hosting.Azure.AppContainers" />
     <PackageReference Include="YamlDotNet" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="../CommunityToolkit.Aspire.Hosting.Dapr/CommunityToolkit.Aspire.Hosting.Dapr.csproj" />
   </ItemGroup>
 
   <ItemGroup>
