+++ src/CommunityToolkit.Aspire.Hosting.Dapr/DaprComponentOptions.cs
