+++ src/CommunityToolkit.Aspire.Hosting.Dapr/DaprComponentReferenceAnnotation.cs
