+++ src/CommunityToolkit.Aspire.Hosting.Dapr/DaprComponentSchema.cs
