+++ src/CommunityToolkit.Aspire.Hosting.Dapr/IDaprComponentResource.cs
